Add image preview widget composable
